refactor: create saner directory structure
