+++ src/cli/commands/destroy/action/runner/destroy-podman-runner.js
@@ -2,7 +2,7 @@ import path from 'path';
 
 import { execa } from 'execa';
 
-import sourceRunnerEnvValues from '../../../utils/source-runner-env-values.js';
+import sourceRunnerEnvValues from '../../../../utils/source-runner-env-values.js';
 
 const destroyDockerRunner = (parsedConfig) =>
   execa(
