+++ src/cli/commands/shell/action/runner/shell-docker-runner.js
@@ -2,18 +2,23 @@ import path from 'path';
 
 import { execa } from 'execa';
 
-import sourceRunnerEnvValues from '../../../utils/source-runner-env-values.js';
+import sourceRunnerEnvValues from '../../../../utils/source-runner-env-values.js';
 
-const shellDockerRunner = (parsedConfig, service) =>
+const shellDockerRunner = (parsedConfig, serviceName) =>
   execa(
-    'podman-compose',
+    'docker',
     [
-      parsedConfig.name ? ['--project-name', parsedConfig.name] : null,
-      ['--file', path.join(process.cwd(), parsedConfig.distDir, 'stack.yml')],
+      'compose',
+      parsedConfig.name
+        ? ['--project-name', parsedConfig.name]
+        : [
+            '--file',
+            path.join(process.cwd(), parsedConfig.distDir, 'stack.yml'),
+          ],
       'exec',
-      service,
+      serviceName,
       ['bash'].concat(
-        service === 'db'
+        serviceName === 'db'
           ? [
               '-c',
               `mysql -u${parsedConfig.environment.db.user} -p${parsedConfig.environment.db.password}`,
