+++ src/cli/commands/shell/action/runner/shell-podman-runner.js
@@ -2,23 +2,18 @@ import path from 'path';
 
 import { execa } from 'execa';
 
-import sourceRunnerEnvValues from '../../../utils/source-runner-env-values.js';
+import sourceRunnerEnvValues from '../../../../utils/source-runner-env-values.js';
 
-const shellDockerRunner = (parsedConfig, serviceName) =>
+const shellDockerRunner = (parsedConfig, service) =>
   execa(
-    'docker',
+    'podman-compose',
     [
-      'compose',
-      parsedConfig.name
-        ? ['--project-name', parsedConfig.name]
-        : [
-            '--file',
-            path.join(process.cwd(), parsedConfig.distDir, 'stack.yml'),
-          ],
+      parsedConfig.name ? ['--project-name', parsedConfig.name] : null,
+      ['--file', path.join(process.cwd(), parsedConfig.distDir, 'stack.yml')],
       'exec',
-      serviceName,
+      service,
       ['bash'].concat(
-        serviceName === 'db'
+        service === 'db'
           ? [
               '-c',
               `mysql -u${parsedConfig.environment.db.user} -p${parsedConfig.environment.db.password}`,
