+++ src/cli/commands/shell/shell.js
@@ -1,7 +1,6 @@
 import { Command, Option } from 'commander';
 
-import shellDockerRunner from './runner/shell-docker-runner.js';
-import shellPodmanRunner from './runner/shell-podman-runner.js';
+import shellActionHandler from './action/index.js';
 
 const buildShellCommand = () => {
   const shell = new Command('shell');
@@ -18,32 +17,7 @@ const buildShellCommand = () => {
         .choices(['wordpress', 'db'])
         .default('wordpress')
     )
-    .action(async function shellActionHandler() {
-      const { config: parsedConfig, service } = this.optsWithGlobals();
-
-      /**
-       * @type {(config: object, service: string) => import('execa').ExecaChildProcess}
-       */
-      let runnerFn;
-
-      switch (parsedConfig.engine) {
-        case 'podman': {
-          runnerFn = shellPodmanRunner;
-          break;
-        }
-        case 'docker': {
-          runnerFn = shellDockerRunner;
-          break;
-        }
-        default: {
-          shell.error(
-            `support for the engine ${parsedConfig.engine} is not implemented just yet`
-          );
-        }
-      }
-
-      runnerFn(parsedConfig, service);
-    });
+    .action(shellActionHandler);
 
   return shell;
 };
