+++ src/cli/commands/start/action/runner/start-docker-runner.js
@@ -2,17 +2,18 @@ import path from 'path';
 
 import { execa } from 'execa';
 
-import sourceRunnerEnvValues from '../../../utils/source-runner-env-values.js';
+import sourceRunnerEnvValues from '../../../../utils/source-runner-env-values.js';
 
-const startPodmanComposeRunner = ({ parsedConfig, detached, verbose }) =>
+const startDockerRunner = ({ parsedConfig, detached, verbose }) =>
   execa(
-    'podman-compose',
+    'docker',
     [
+      'compose',
       parsedConfig.name ? ['--project-name', parsedConfig.name] : null,
       ['--file', path.join(process.cwd(), parsedConfig.distDir, 'stack.yml')],
       'up',
       [parsedConfig.environment.rebuildOnStart ? '--build' : null],
-      [detached ? '--detach' : '--abort-on-container-exit'],
+      [detached ? '--detached' : null],
       [verbose ? null : '--quiet-pull'],
     ]
       .flat()
@@ -22,4 +23,4 @@ const startPodmanComposeRunner = ({ parsedConfig, detached, verbose }) =>
     }
   );
 
-export default startPodmanComposeRunner;
+export default startDockerRunner;
