+++ src/cli/commands/start/start.js
@@ -1,16 +1,9 @@
-import path from 'path';
-import * as url from 'url';
 import { createRequire } from 'module';
 
-import cpy from 'cpy';
 import Listr from 'listr';
 import { Command, Option } from 'commander';
-import { writeJsonFile } from 'write-json-file';
 
-import generateComposerConfig from '../../utils/generate-composer-config.js';
-
-import startDockerRunner from './runner/start-docker-runner.js';
-import startPodmanComposeRunner from './runner/start-podman-compose-runner.js';
+import startActionHandler from './action/index.js';
 
 const requireFn = createRequire(import.meta.url);
 
@@ -54,78 +47,7 @@ const buildStartCommand = () => {
         command.error(e.message);
       }
     })
-    .action(async function startActionHandler() {
-      const {
-        config: parsedConfig,
-        detached,
-        verbose,
-      } = this.optsWithGlobals();
-
-      await Promise.allSettled([
-        cpy(
-          path.join(
-            path.dirname(url.fileURLToPath(import.meta.url)),
-            '../../../templates/core/*'
-          ),
-          path.join(process.cwd(), parsedConfig.distDir)
-        ),
-        writeJsonFile(
-          path.join(process.cwd(), parsedConfig.distDir, 'composer.json'),
-          generateComposerConfig(parsedConfig.wpackagist),
-          {
-            indent: 2,
-          }
-        ),
-      ]);
-
-      /**
-       * @type {(config: object) => import('execa').ExecaChildProcess}
-       */
-      let runnerFn;
-
-      switch (parsedConfig.engine) {
-        case 'podman': {
-          runnerFn = startPodmanComposeRunner;
-          break;
-        }
-        case 'docker': {
-          runnerFn = startDockerRunner;
-          break;
-        }
-        default: {
-          start.error(
-            `support for the engine ${parsedConfig.engine} is not implemented just yet`
-          );
-        }
-      }
-
-      const runner = runnerFn({
-        parsedConfig,
-        detached,
-        verbose,
-      });
-
-      // setup handler to terminate runner using CTRL+C
-      process.on(
-        'SIGINT',
-        () =>
-          new Promise((resolve, reject) => {
-            // send signal to cancel to our runner
-            runner.cancel();
-
-            runner.once('exit', () => {
-              resolve(undefined);
-            });
-
-            runner.once('error', (err) => {
-              reject(err);
-            });
-          })
-      );
-
-      runner.stdout.pipe(process.stdout);
-      runner.stderr.pipe(process.stderr);
-    });
+    .action(startActionHandler);
 
   return start;
 };
