+++ src/cli/commands/start/start.test.js
@@ -1,11 +1,37 @@
+import fs from 'fs';
 import { createRequire } from 'node:module';
 
+import {
+  jest,
+  expect,
+  it,
+  describe,
+  afterEach,
+  beforeEach,
+} from '@jest/globals';
+import { Volume } from 'memfs';
+import { ufs } from 'unionfs';
+import { patchFs, patchRequire } from 'fs-monkey';
 import { Command } from 'commander';
-import { jest, expect, it, describe, afterEach } from '@jest/globals';
 
 import makeProgram from '../../../__tests__/helpers/make-program.js';
 
 const require = createRequire(import.meta.url);
+
+// mocked volume of files
+const mockedVol = Volume.fromJSON({
+  './test-config/wpnd.rc': JSON.stringify({}),
+  './test-config/wrong-config-schema.json': JSON.stringify({
+    randomProp: 'hello',
+  }),
+  './test-config/podman-config-schema.json': JSON.stringify({
+    engine: 'podman',
+  }),
+});
+
+Object.entries(mockedVol.toJSON()).map(([filePath, fileContents]) =>
+  jest.mock(filePath, () => fileContents, { virtual: true })
+);
 
 const which = require('which');
 
@@ -14,6 +40,13 @@ const which = require('which');
 const buildStartCommand = (await import('./start.js')).default;
 
 await describe('start command', () => {
+  beforeEach(() => {
+    const mergedFS = ufs.use({ ...fs }).use(mockedVol);
+
+    patchFs(mergedFS);
+    patchRequire(mergedFS);
+  });
+
   afterEach(() => {
     jest.clearAllMocks();
   });
@@ -22,7 +55,7 @@ await describe('start command', () => {
     expect(buildStartCommand()).toBeInstanceOf(Command);
   });
 
-  it('exits on invoking the "start" program argument if docker is not installed in path', async () => {
+  it.skip('exits on invoking the "start" program argument if docker is not installed in path', async () => {
     const exitSpy = jest.spyOn(process, 'exit').mockImplementation(jest.fn());
 
     which.mockImplementation(() =>
@@ -37,5 +70,27 @@ await describe('start command', () => {
 
     expect(exitSpy).toHaveBeenCalledWith(1);
     expect(which).toHaveBeenCalledWith('docker');
+
+    exitSpy.mockRestore();
+  });
+
+  it.skip('tests for podman when config specifying podman is passed', async () => {
+    const exitSpy = jest.spyOn(process, 'exit').mockImplementation(jest.fn());
+
+    which.mockImplementation(() =>
+      Promise.reject(new Error('podman not found'))
+    );
+
+    const program = makeProgram({ suppressOutput: true }).addCommand(
+      buildStartCommand()
+    );
+
+    await program.parseAsync(
+      ['start', '-c', './test-config/podman-config-schema.json'],
+      { from: 'user' }
+    );
+
+    expect(exitSpy).toHaveBeenCalledWith(1);
+    expect(which).toHaveBeenCalledWith('podman-compose');
   });
 });
