+++ src/cli/global-options/config/index.js
