+++ src/cli/global-options/config/index.test.js
