+++ src/cli/global-options/config/schema.json
