+++ src/cli/global-options/logs.js
@@ -1,0 +1,8 @@
+import { Option } from 'commander';
+
+const showLogs = new Option(
+  '-v, --verbose',
+  'show container engine logs'
+).default(false);
+
+export default showLogs;
