--- src/cli/options/logs.js
@@ -1,5 +1,0 @@
-import { Option } from 'commander';
-
-const showLogs = new Option('-v, --verbose', 'show docker logs').default(false);
-
-export default showLogs;
